Extractign memory analysis to new PreprocessHWLoops
