+++ apps/hardware_benchmarks/tests/merged_unit_tests/test_main.cpp
@@ -2599,140 +2599,6 @@ void real_unsharp_test() {
   //assert(false);
 }
 
-class MemOp {
-  public:
-
-    std::vector<const For*> surroundingLoops;
-    const Store* store;
-    const Load* load;
-
-    MemOp() : store(nullptr), load(nullptr) {}
-    MemOp(std::vector<const For*>& lps_,
-        const Store* st_,
-        const Load* ld_) : surroundingLoops(lps_), store(st_), load(ld_) {}
-
-    std::string getBufferName() const {
-      if (isLoad()) {
-        return load->name;
-      }
-      return store->name;
-    }
-
-    bool isLoad() const { return load != nullptr; }
-    bool isStore() const { return store != nullptr; }
-    
-
-    void print(std::ostream& out) const {
-      for (auto lp : surroundingLoops) {
-        out << lp->name << " : " << lp->min << ", " << lp->extent << "; ";
-      }
-      if (isLoad()) {
-        out << "ld: " << load->name << "[" << load->index << "]";
-      } else {
-        assert(isStore());
-        //out << "st: " << store->name << "[" << store->index << "] =  " << store->value;
-        out << "st: " << store->name << "[" << store->index << "]";
-      }
-    }
-};
-
-std::ostream& operator<<(std::ostream& out, const MemOp& op) {
-  op.print(out);
-  return out;
-}
-
-class MemoryInfoCollector : public IRGraphVisitor {
-  public:
-
-    using IRGraphVisitor::visit;
-
-    std::vector<const For*> activeLoops;
-    std::vector<MemOp> memOps;
-    std::map<string, Expr> letValues;
-
-    std::set<std::string> allBufferNames() const {
-      set<string> names;
-      for (auto op : memOps) {
-        names.insert(op.getBufferName());
-      }
-
-      return names;
-    }
-
-    std::vector<MemOp> allOpsOn(const std::string& buf) const {
-      vector<MemOp> ops;
-      for (auto op : memOps) {
-        if (op.getBufferName() == buf) {
-          ops.push_back(op);
-        }
-      }
-      return ops;
-    }
-
-    std::set<string> roms() const {
-      // Find all memories that are only stored to at constant indexes outside of loop nests
-      set<string> memNames = allBufferNames();
-      set<string> roms;
-      for (auto buf : memNames) {
-        vector<MemOp> ops = allOpsOn(buf);
-        bool allStoresToConstants = true;
-        int numStores = 0;
-        for (auto op : ops) {
-          if (op.isStore()) {
-            numStores++;
-            const Expr& ind = op.store->index;
-            if (!Halide::Internal::is_const(ind) ||
-                !Halide::Internal::is_const(op.store->value)) {
-              allStoresToConstants = false;
-              break;
-            }
-          }
-        }
-        if (numStores > 0 && allStoresToConstants) {
-          roms.insert(buf);
-        }
-      }
-      return roms;
-    }
-
-    void visit(const Let* lt) {
-      letValues[lt->name] = lt->value;
-      IRGraphVisitor::visit(lt);
-      letValues.erase(lt->name);
-    }
-
-    void visit(const LetStmt* lt) {
-      letValues[lt->name] = lt->value;
-      IRGraphVisitor::visit(lt);
-      letValues.erase(lt->name);
-    }
-
-    void visit(const Load* ld) {
-      IRGraphVisitor::visit(ld);
-      Expr addr = substitute(letValues, ld->index);
-
-      cout << "Old = " << ld->index << ", New addr = " << addr << endl;
-
-      memOps.push_back({activeLoops, nullptr, ld});
-    }
-
-    void visit(const Store* st) {
-      IRGraphVisitor::visit(st);
-      Expr addr = substitute(letValues, st->index);
-      memOps.push_back({activeLoops, st, nullptr});
-    }
-
-    void visit(const For* lp) {
-      activeLoops.push_back(lp);
-
-
-      IRGraphVisitor::visit(lp);
-
-      activeLoops.pop_back();
-    }
-
-};
-
 void conv_layer_mobile_test() {
   ImageParam input(type_of<int8_t>(), 3);
   ImageParam output(type_of<int8_t>(), 3);
